split apart archive.py to archive and tar_writer
No behavior change. Just make the division cleaner.
